test: split CurrencyTest into one class per production class
